(BSR)[PRO] refactor: Header with new archi, delete unused files
