+++ pro/src/app/App/hook/__specs__/useLogExtraProData.spec.tsx
@@ -7,7 +7,7 @@ import * as useAnalytics from 'app/App/analytics/firebase'
 import { useLogExtraProData } from 'app/App/hook/useLogExtraProData'
 import { getOffererNameFactory } from 'commons/utils/factories/individualApiFactories'
 import { renderWithProviders } from 'commons/utils/renderWithProviders'
-import { HeaderDropdown } from 'components/Header/HeaderDropdown/HeaderDropdown'
+import { HeaderDropdown } from 'components/Header/components/HeaderDropdown/HeaderDropdown'
 
 const mockLogEvent = vi.fn()
