+++ pro/src/components/ActionsBarSticky/ActionsBarSticky.spec.tsx
@@ -3,7 +3,7 @@ import { screen } from '@testing-library/react'
 import { Mode } from 'commons/core/OfferEducational/types'
 import { renderWithProviders } from 'commons/utils/renderWithProviders'
 
-import { ActionsBarSticky } from '../ActionsBarSticky'
+import { ActionsBarSticky } from './ActionsBarSticky'
 
 const renderActionsBar = ({
   dirtyForm,
