--- pro/src/components/Header/OldHeader.module.scss
@@ -1,84 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/variables/_size.scss" as size;
-
-.menu-v2 {
-  background-image: var(--color-linear-gradient-to-right);
-  min-height: rem.torem(160px);
-
-  .nav {
-    display: grid;
-    grid-template-columns: size.$header-grid-template-columns;
-
-    .nav-brand {
-      display: flex;
-      align-items: center;
-      justify-self: end;
-      margin-right: size.$nav-brand-margin-right;
-    }
-
-    .nav-menu {
-      display: flex;
-      height: rem.torem(72px);
-      justify-content: center;
-
-      .icon-only.nav-item {
-        background: none;
-        border: none;
-      }
-
-      .nav-item {
-        @include fonts.body-exergue;
-
-        align-items: center;
-        color: var(--color-white);
-        display: flex;
-        height: 100%;
-        padding: 0 size.$header-nav-item-padding;
-
-        .nav-item-icon {
-          height: rem.torem(24px);
-          margin-right: rem.torem(8px);
-          width: rem.torem(24px);
-          fill: white;
-        }
-
-        &.icon-only {
-          flex-shrink: 0;
-          padding: 0 rem.torem(8px);
-
-          .signout-icon {
-            margin-right: 0;
-          }
-        }
-
-        &:hover,
-        &:focus,
-        &-selected {
-          background-color: var(--color-primary);
-        }
-      }
-
-      & :last-child {
-        display: flex;
-      }
-
-      .separator {
-        align-self: center;
-        background-color: var(--color-white);
-        flex-shrink: 0;
-        height: rem.torem(24px);
-        width: rem.torem(1px);
-      }
-    }
-  }
-}
-
-.logo {
-  align-items: center;
-  color: var(--color-white);
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: center;
-}
